Create Encryption 101
THM Write Up 2. In this Write-up, we went over encryption and briefly touch the surfaces of the types of encryption and why it’s very important.  Encryption is important because this is how we can verify when we connect to our bank and are not being scammed. The two main categories of Encryption are symmetric and asymmetric. In symmetric encryption, the same key is used for encryption and decryption. Asymmetric encryption A different key is used for the encryption and decryption process. One of the keys is typically known as the private key and the other is known as the public key. HTTPS is a popularly known type of encryption. Another way encryption works are through digital signatures and certificates. This is how a website can trust one another and make sure it’s the real company that you are communicating to.” SSL certificates have an associated public/private key pair. The public key is distributed as part of the certificate, and the private key is kept incredibly safely guarded. “ Encryptions uses  SSH private keys which should be treated like passwords.